🚚 Move files to src folder
