+++ src/__tests__/OtpInput.test.js
@@ -2,7 +2,7 @@ import { TextInput, Platform } from 'react-native'
 import React from 'react'
 import renderer from 'react-test-renderer'
 
-import OtpInput from '../lib/OtpInput'
+import OtpInput from '../OtpInput'
 
 jest.mock('Platform', () => ({
   OS: 'ios',
